Project Development Phase/Sprint 4/requirements.txt
blinker==1.5
click==8.1.3
Flask==2.2.2
Flask-DB2==0.0.10
Flask-Mail==0.9.1
Flask-WTF==1.0.1
ibm-db==3.1.3
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
plotly==5.11.0
python-dotenv==0.21.0
tenacity==8.1.0
timeago==1.0.16
Werkzeug==2.2.2
WTForms==2.2.1
sendgrid==6.9.7
